Move interface classes to correct directory
Interfaces that used in several components moved to src/components/include/<component_name>
Mocks of this interfaces moved to src/components/include/test/<component_name>

Related tasks APPLINK-30588 APPLINK-30974
