+++ src/components/include/application_manager/application_manager.h
@@ -36,7 +36,7 @@
 #include <string>
 #include <vector>
 #include <set>
-#include "vehicle_info_data.h"
+#include "application_manager/vehicle_info_data.h"
 #include "application_manager/application.h"
 #include "application_manager/hmi_capabilities.h"
 #include "application_manager/commands/command.h"
