+++ src/components/include/application_manager/policies/policy_handler_observer.h
