+++ src/components/include/application_manager/vehicle_info_data.h
