+++ src/components/include/hmi_message_handler/hmi_message_adapter.h
